# apps/intake-api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Votatis Intake API",
    "version": "1.0.0"
  },
  "components": {
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ]
      },
      "Region": {
        "type": "object",
        "properties": {
          "sido": {
            "type": "string"
          },
          "sigungu": {
            "type": "string"
          },
          "eup_myeon_dong": {
            "type": "string"
          }
        }
      },
      "Source": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string",
            "format": "uri"
          },
          "text": {
            "type": "string",
            "minLength": 1
          },
          "type": {
            "type": "string"
          },
          "captured_at": {
            "type": "string"
          },
          "archive_url": {
            "type": "string",
            "format": "uri"
          }
        }
      },
...
      "Attachment": {
        "type": "object",
        "properties": {
          "filename": {
            "type": "string"
          },
          "r2_key": {
            "type": "string"
          },
          "sha256": {
            "type": "string"
          },
          "mime": {
            "type": "string"
          },
          "size": {
            "type": "integer"
          }
        },
        "required": [
          "filename",
          "r2_key",
          "sha256",
          "mime",
          "size"
        ]
      },
...
      "ReportSummary": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "election": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "summary": {
            "type": [
              "string",
              "null"
            ]
          },
          "region": {
            "$ref": "#/components/schemas/Region"
          },
          "occurred_at": {
            "type": [
              "string",
              "null"
            ]
          },
          "collected_at": {
            "type": "string"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "attachment_count": {
            "type": "integer"
          }
        },
        "required": [
          "id",
          "status",
          "election",
          "title",
          "summary",
          "region",
          "occurred_at",
          "collected_at",
          "tags",
          "attachment_count"
        ]
      },
...
      "Verification": {
        "type": "object",
        "properties": {
          "reviewer": {
            "type": [
              "string",
              "null"
            ]
          },
          "method": {
            "type": [
              "string",
              "null"
            ]
          },
          "reviewed_at": {
            "type": [
              "string",
              "null"
            ]
          },
          "notes": {
            "type": [
              "string",
              "null"
            ]
          },
          "evidence_links": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "reviewer",
          "method",
          "reviewed_at",
          "notes",
          "evidence_links"
        ]
      },
      "Report": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "election": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "summary": {
            "type": [
              "string",
              "null"
            ]
          },
          "body": {
            "type": [
              "string",
              "null"
            ]
          },
          "region": {
            "$ref": "#/components/schemas/Region"
          },
          "occurred_at": {
            "type": [
              "string",
              "null"
            ]
          },
          "collected_at": {
            "type": "string"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "sources": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Source"
            }
          },
          "attachments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Attachment"
            }
          },
          "rebuttals": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "object",
              "properties": {
                "text": {
                  "type": "string"
                },
                "source_url": {
                  "type": "string"
                }
              },
              "required": [
                "text"
              ]
            }
          },
          "related": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            }
          },
          "consent": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "license": {
            "type": "string"
          },
          "verification": {
            "$ref": "#/components/schemas/Verification"
          },
          "created_at": {
            "type": "string"
          },
          "updated_at": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "status",
          "election",
          "title",
          "summary",
          "body",
          "region",
          "occurred_at",
          "collected_at",
          "tags",
          "sources",
          "attachments",
          "rebuttals",
          "related",
          "consent",
          "license",
          "verification",
          "created_at",
          "updated_at"
        ]
      },
      "AdminSessionResult": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "enum": [
              true
            ]
          }
        },
        "required": [
          "ok"
        ]
      },
      "AdminSessionInput": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "token"
        ]
      },
      "AdminReportSummary": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ReportSummary"
          },
          {
            "type": "object",
            "properties": {
              "submitter": {
                "type": "string"
              },
              "updated_at": {
                "type": "string"
              }
            },
            "required": [
              "submitter",
              "updated_at"
            ]
          }
        ]
      },
      "AdminReportList": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AdminReportSummary"
            }
          },
          "total": {
            "type": "integer"
          },
          "counts": {
            "type": "object",
            "additionalProperties": {
              "type": "integer"
            }
          },
          "limit": {
            "type": "integer"
          },
          "offset": {
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "counts",
          "limit",
          "offset"
        ]
      },
      "AdminReportDetail": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Report"
          },
          {
            "type": "object",
            "properties": {
              "submitter": {
                "type": "string"
              },
              "exif": {
                "type": [
                  "array",
                  "null"
                ],
                "items": {}
              }
            },
            "required": [
              "submitter",
              "exif"
            ]
          }
        ]
      },
      "AdminPatch": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "unverified",
              "reviewing",
              "confirmed",
              "disputed",
              "debunked",
              "corrected"
            ]
          },
          "reviewer": {
            "type": "string"
          },
          "verification": {
            "type": "object",
            "properties": {
              "reviewer": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "method": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "notes": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "evidence_links": {
                "type": "array",
                "items": {
                  "type": "string",
                  "format": "uri"
                }
              }
            }
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "maxItems": 50
          },
          "rebuttals": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "text": {
                  "type": "string",
                  "minLength": 1
                },
                "source_url": {
                  "type": "string",
                  "format": "uri"
                }
              },
              "required": [
                "text"
              ]
            }
          },
          "related": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "title": {
            "type": "string",
            "minLength": 1
          },
          "summary": {
            "type": "string"
          },
          "body": {
            "type": "string"
          }
        }
      },
      "Analysis": {
        "type": "object",
        "properties": {
          "suggested_tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "credibility": {
            "type": "object",
            "properties": {
              "score": {
                "type": "number"
              },
              "signals": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            },
            "required": [
              "score",
              "signals"
            ]
          },
          "synthetic_risk": {
            "type": "object",
            "properties": {
              "level": {
                "type": "string",
                "enum": [
                  "low",
                  "review",
                  "unknown"
                ]
              },
              "signals": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            },
            "required": [
              "level",
              "signals"
            ]
          },
          "summary_hint": {
            "type": "string"
          },
          "source": {
            "type": "string",
            "enum": [
              "heuristic",
              "heuristic+ai"
            ]
          }
        },
        "required": [
          "suggested_tags",
          "credibility",
          "synthetic_risk",
          "summary_hint",
          "source"
        ]
      },
      "AdminExport": {
        "type": "object",
        "properties": {
          "records": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Report"
            }
          }
        },
        "required": [
          "records"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/admin/session": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminSessionInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "토큰 유효",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminSessionResult"
                }
              }
            }
          },
          "401": {
            "description": "토큰 불일치",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/admin/reports": {
      "get": {
        "parameters": [
          {
            "schema": {
              "type": "string",
              "enum": [
                "unverified",
                "reviewing",
                "confirmed",
                "disputed",
                "debunked",
                "corrected"
              ]
            },
            "required": false,
            "name": "status",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "election",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "sido",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "sigungu",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "tag",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "q",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0,
              "maximum": 100,
              "default": 20
            },
            "required": false,
            "name": "limit",
            "in": "query"
          },
          {
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "minimum": 0,
              "default": 0
            },
            "required": false,
            "name": "offset",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "검수 큐 목록 + 상태별 카운트",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminReportList"
                }
              }
            }
          },
          "401": {
            "description": "인증 필요",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/admin/reports/{id}": {
      "get": {
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "상세(내부 필드 포함)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminReportDetail"
                }
              }
            }
          },
          "401": {
            "description": "인증 필요",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "patch": {
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "갱신된 상세",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminReportDetail"
                }
              }
            }
          },
          "400": {
            "description": "근거 누락 등 검증 실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "인증 필요",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/admin/reports/{id}/analyze": {
      "post": {
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "보조 분석 신호(휴리스틱±AI)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Analysis"
                }
              }
            }
          },
          "401": {
            "description": "인증 필요",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/admin/export": {
      "get": {
        "responses": {
          "200": {
            "description": "검증 완료 레코드(공개 필드)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminExport"
                }
              }
            }
          },
          "401": {
            "description": "인증 필요",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}